% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{calculateLandscapeMetrics}
\alias{calculateLandscapeMetrics}
\title{Calculate landscape metrics}
\usage{
calculateLandscapeMetrics(summaryPolys, polyCol, vtm)
}
\arguments{
\item{summaryPolys}{\code{sf} polygons object}

\item{polyCol}{character, specifying the polygon attribute (column) in \code{summaryPolys} to use}

\item{vtm}{character, specifying the file path to a vegetation type map raster (\code{.tif})}
}
\value{
summary \code{data.frame} object
}
\description{
Calculate landscape metrics
}
