% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{calculatePatchMetrics}
\alias{calculatePatchMetrics}
\title{Calculate patch statistics/metrics}
\usage{
calculatePatchMetrics(summaryPolys, polyCol, flm, vtm, sam, funList = NULL)
}
\arguments{
\item{summaryPolys}{\code{sf} polygons object}

\item{polyCol}{character, specifying the polygon attribute (column) in \code{summaryPolys} to use}

\item{flm}{character, specifying the file path to a flammability map raster (\code{.tif})}

\item{vtm}{character, specifying the file path to a vegetation type map raster (\code{.tif})}

\item{sam}{character, specifying the file path to a stand age map raster (\code{.tif})}

\item{funList}{list of character strings specifying function names used to calculate summary
patch metrics. When \code{NULL}, \code{\link[=default_patch_metrics]{default_patch_metrics()}} is used.}
}
\value{
summary \code{data.frame} object
}
\description{
Calculate patch statistics/metrics
}
\seealso{
\code{\link[=calculatePatchMetricsSeral]{calculatePatchMetricsSeral()}}
}
