% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-stats.R
\name{calculatePatchMetricsSeral}
\alias{calculatePatchMetricsSeral}
\title{Calculate patch statistics/metrics}
\usage{
calculatePatchMetricsSeral(summaryPolys, polyCol, flm, ssm)
}
\arguments{
\item{summaryPolys}{\code{sf} polygons object}

\item{polyCol}{character, specifying the polygon attribute (column) in \code{summaryPolys} to use}

\item{flm}{character, specifying the file path to a flammability map raster (\code{.tif})}

\item{ssm}{character, specifying the file path to a seral stage map raster (\code{.tif})}
}
\value{
summary \code{data.frame} object
}
\description{
Calculate patch statistics/metrics
}
\seealso{
\code{\link[=calculatePatchMetrics]{calculatePatchMetrics()}}
}
