% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegcover.R
\docType{data}
\name{.ageClassCutOffs}
\alias{.ageClassCutOffs}
\title{Age class cut-offs (years)}
\format{
An object of class \code{integer} of length 4.
}
\usage{
.ageClassCutOffs
}
\description{
Age class cut-offs (years)
}
\keyword{internal}
