% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegcover.R
\docType{data}
\name{.ageClasses}
\alias{.ageClasses}
\title{Age class names}
\format{
An object of class \code{character} of length 4.
}
\usage{
.ageClasses
}
\description{
Age class names
}
\keyword{internal}
