% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecodistricts}
\alias{ecodistricts}
\title{Terrestrial Ecodistricts of Canada}
\format{
\subsection{\code{ecodistricts}}{

An \code{sf} polygons object with 1025 features.
\describe{
\item{OBJECTID, SHAPE_Length, SHAPE_Area}{System-generated attributes}
\item{ECODISTRICT_ID}{Ecodistrict identifier}
\item{ECOREGION_ID}{Ecoregion identifier}
\item{ECOZONE_ID}{Ecozone identifier}
\item{ECOPROVINCE_ID}{Ecoprovince identifier}
\item{geometry}{\code{sfc_POLYGON}}
}
}
}
\source{
\url{https://open.canada.ca/data/en/dataset/fe9fd41c-1f67-4bc5-809d-05b62986b26b}
}
\usage{
ecodistricts
}
\description{
Terrestrial Ecodistricts of Canada
}
\keyword{datasets}
