% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoprovinces}
\alias{ecoprovinces}
\title{Terrestrial Ecoprovinces of Canada}
\format{
\subsection{\code{ecoprovinces}}{

An \code{sf} polygons object with 68 features.
\describe{
\item{OBJECTID, SHAPE_Length, SHAPE_Area}{System-generated attributes}
\item{ECOZONE_ID}{Ecozone identifier}
\item{ECOPROVINCE_ID}{Ecoprovince identifier}
\item{ECOPROVINCE_NAME_EN}{English ecoprovince name}
\item{ECOPROVINCE_NAME_FR}{French ecoprovince name}
\item{geometry}{\code{sfc_POLYGON}}
}
}
}
\source{
\url{https://open.canada.ca/data/en/dataset/98fa7335-fbfe-4289-9a0e-d6bf3874b424}
}
\usage{
ecoprovinces
}
\description{
Terrestrial Ecoprovinces of Canada
}
\keyword{datasets}
