% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Terrestrial Ecoregions of Canada}
\format{
\subsection{\code{ecoregions}}{

An \code{sf} polygons object with 218 features.
\describe{
\item{OBJECTID, SHAPE_Length, SHAPE_Area}{System-generated attributes}
\item{ECOREGION_ID}{Ecoregion identifier}
\item{ECOZONE_ID}{Ecozone identifier}
\item{ECOPROVINCE_ID}{Ecoprovince identifier}
\item{ECOREGION_NAME_EN}{English ecoregion name}
\item{ECOREGION_NAME_FR}{French ecoregion name}
\item{geometry}{\code{sfc_POLYGON}}
}
}
}
\source{
\url{https://open.canada.ca/data/en/dataset/ade80d26-61f5-439e-8966-73b352811fe6}
}
\usage{
ecoregions
}
\description{
Terrestrial Ecoregions of Canada
}
\keyword{datasets}
