% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecozones}
\alias{ecozones}
\title{Terrestrial Ecozones of Canada}
\format{
\subsection{\code{ecozones}}{

An \code{sf} polygons object with 25 features.
\describe{
\item{OBJECTID, SHAPE_Length, SHAPE_Area}{System-generated attributes}
\item{ECOZONE_ID}{Ecozone identifier}
\item{ECOZONE_NAME_EN}{English ecozone name}
\item{ECOZONE_NAME_FR}{French ecozone name}
\item{geometry}{\code{sfc_POLYGON}}
}
}
}
\source{
\url{https://open.canada.ca/data/en/dataset/7ad7ea01-eb23-4824-bccc-66adb7c5bdf8}
}
\usage{
ecozones
}
\description{
Terrestrial Ecozones of Canada
}
\keyword{datasets}
