% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{patchAges}
\alias{patchAges}
\title{Calculate median stand age for each patch (per species)}
\usage{
patchAges(vtm, sam)
}
\arguments{
\item{vtm}{character, specifying the file path to a vegetation type map raster (\code{.tif})}

\item{sam}{character, specifying the file path to a stand age map raster (\code{.tif})}
}
\value{
\code{data.table} with columns \code{layer}, \code{level}, \code{class}, \code{id}, \code{metric},
and \code{value}, where \code{class} contains species names and \code{value} is the median
stand age (years) for each patch.
}
\description{
Calculate median stand age for each patch (per species)
}
