% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{patchStats}
\alias{patchStats}
\title{Calculate patch statistics/metrics}
\usage{
patchStats(vtm, sam, flm, polyNames, summaryPolys, polyCol, funList)
}
\arguments{
\item{vtm}{character, specifying the file path to a vegetation type map raster (\code{.tif})}

\item{sam}{character, specifying the file path to a stand age map raster (\code{.tif})}

\item{flm}{character, specifying the file path to a flammability map raster (\code{.tif})}

\item{polyNames}{character, specifying polygon names in \code{summaryPolys} over which
to summarize patch statistics}

\item{summaryPolys}{\code{sf} polygons object}

\item{polyCol}{character, specifying the polygon attribute (column) in \code{summaryPolys} to use}

\item{funList}{list of character strings specifying functions to use for calculating
patch statistics (e.g., \code{list("landscapemetrics::lsm_l_area_mn")}).
These functions must work analogously to those in \pkg{landscapemetrics}.}
}
\value{
nested list of summary data.frames containing patch statistics by summary polygon
}
\description{
Calculate patch statistics/metrics
}
\seealso{
\code{\link[=patchStatsSeral]{patchStatsSeral()}}
}
