% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{summarizePatchMetrics}
\alias{summarizePatchMetrics}
\alias{summarizePatchMetricsSeral}
\title{Summarize patch statistics/metrics}
\usage{
summarizePatchMetrics(ptch_stat_df)

summarizePatchMetricsSeral(ptch_stat_df)
}
\arguments{
\item{ptch_stat_df}{named list of patch stat \code{data.frame} objects,
with names corresponding to those of \code{funList} passed to
\code{calculatePatchMetrics()} or \code{calculatePatchMetricsSeral()}.}
}
\value{
summary \code{data.frame} object
}
\description{
Summarize patch statistics/metrics
}
