% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{calculateLandscapeMetrics}
\alias{calculateLandscapeMetrics}
\title{Calculate landscape metrics}
\usage{
calculateLandscapeMetrics(summaryPolys, polyCol, vtm, funList = NULL)
}
\arguments{
\item{summaryPolys}{\code{sf} polygons object}

\item{polyCol}{character, specifying the polygon attribute (column) in \code{summaryPolys} to use}

\item{vtm}{character, specifying the file path to a vegetation type map raster (\code{.tif})}

\item{funList}{optional list of character strings specifying functions to use for calculating
landscape or patch statistics (e.g., \code{list("lsm_l_area_mn")}).
These functions must work analogously to those in \pkg{landscapemetrics}.
If not specified by the user (i.e., \code{funList = NULL}), default function names are used:
see \code{\link[=default_landscape_metrics]{default_landscape_metrics()}}, \code{\link[=default_patch_metrics]{default_patch_metrics()}}, \code{\link[=default_patch_metrics_seral]{default_patch_metrics_seral()}}.}
}
\value{
summary \code{data.frame} object
}
\description{
Calculate landscape metrics
}
