% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-metrics.R
\name{default_landscape_metrics}
\alias{default_landscape_metrics}
\alias{default_patch_metrics}
\alias{default_patch_metrics_seral}
\title{Default metrics}
\usage{
default_landscape_metrics()

default_patch_metrics()

default_patch_metrics_seral()
}
\description{
Default metrics
}
\note{
patch ages don't make sense here since seral stage (based on age) determines patches
}
\section{Default Landscape Metrics}{


The following landscape-level metrics are calculated using \pkg{landscapemetrics}:
\itemize{
\item \code{lsm_l_area_mn}: mean patch area;
\item \code{lsm_l_area_sd}: standard deviation of patch area;
\item \code{lsm_l_cohesion}, ## cohesion index;
\item \code{lsm_l_condent}: conditional entropy;
\item \code{lsm_l_core_cv}: coeff var of core area;
\item \code{lsm_l_core_mn}: mean core area;
\item \code{lsm_l_core_sd}: std dev of core area;
\item \code{lsm_l_ed}: edge density;
\item \code{lsm_l_iji}: interspersion and juxtaposition index;
}
}

\section{Default Patch Metrics}{


The following patch- and class-level metrics are calculated using
\pkg{landscapemetrics} and \pkg{nrvtools}:
\itemize{
\item \code{patchAges}: patch ages;
\item \code{patchAreas}: patch area (uses lsm_p_area);
\item \code{lsm_c_ai}: aggregation index by class;
\item \code{lsm_c_area_cv}: coeff var patch area by class;
\item \code{lsm_c_area_mn}: mean patch area by class;
\item \code{lsm_c_area_sd}: std dev patch area by class;
\item \code{lsm_c_ca}: total class area;
}
}

\section{Default Seral Stage Patch Metrics}{


The following patch- and class-level metrics are calculated using
\pkg{landscapemetrics} and \pkg{nrvtools}:
\itemize{
\item \code{patchAreasSeral}:  patch area (uses lsm_p_area);
\item the subset of \code{default_patch_metrics()} from \pkg{landscapemetrics}
(i.e., prefixed with 'lsm_')
}
}

