% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seral-stages.R
\docType{data}
\name{.seralStagesBC}
\alias{.seralStagesBC}
\title{Seral stage classes for BC forests}
\format{
An object of class \code{character} of length 16.
}
\usage{
.seralStagesBC
}
\description{
"early", "mid", "mature", and "old" are for NDT1, NDT2, NDT3;
"early_Fir", "mid_Fir", "mature_Fir", and "old_Fir" are for NDT4 with Fir leading;
"early_Pine", "mid_Pine", "mature_Pine", and "old_Pine" are for NDT4 with Pine leading;
"early_Other", "mid_Other", "mature_Other", and "old_Other" are for NDT4 with Other leading.
}
\keyword{internal}
