% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrvtools-package.R
\docType{package}
\name{nrvtools-package}
\alias{nrvtools}
\alias{nrvtools-package}
\title{nrvtools: Tools for Summarizing Outputs From Natural Range of Variability Simulations}
\description{
Tools and utilities for working with outputs from NRV simulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FOR-CAST/nrvtools}
  \item Report bugs at \url{https://github.com/FOR-CAST/nrvtools/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

}
\keyword{internal}
