% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-stats.R
\name{patchAges}
\alias{patchAges}
\title{Calculate median stand age for each patch (per species)}
\usage{
patchAges(vtm, sam)
}
\arguments{
\item{vtm}{character, specifying the file path to a vegetation type map raster (\code{.tif})}

\item{sam}{character, specifying the file path to a stand age map raster (\code{.tif})}
}
\value{
\code{tibble} object from \code{landscapemetrics::lsm_p_area}
}
\description{
Calculate median stand age for each patch (per species)
}
