% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-stats.R
\name{patchAreas}
\alias{patchAreas}
\title{Calculate areas for each patch (per species)}
\usage{
patchAreas(vtm)
}
\arguments{
\item{vtm}{character, specifying the file path to a vegetation type map raster (\code{.tif})}
}
\value{
\code{tibble} object from \code{landscapemetrics::lsm_p_area}
}
\description{
Calculate areas for each patch (per species)
}
\seealso{
\code{\link[=patchAreasSeral]{patchAreasSeral()}}
}
