% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{patchAreasSeral}
\alias{patchAreasSeral}
\title{Calculate median stand age for each patch (by seral stage)}
\usage{
patchAreasSeral(ssm)
}
\arguments{
\item{ssm}{character, specifying the file path to a seral stage map raster (\code{.tif})}
}
\value{
\code{tibble} object from \code{landscapemetrics::lsm_p_area}
}
\description{
Calculate median stand age for each patch (by seral stage)
}
\seealso{
\code{\link[=patchAreas]{patchAreas()}}
}
