% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{patchStatsSeral}
\alias{patchStatsSeral}
\title{Calculate patch statistics (metrics) based on seral stages}
\usage{
patchStatsSeral(ssm, flm, polyNames, summaryPolys, polyCol, funList)
}
\arguments{
\item{ssm}{character, specifying the file path to a seral stage map raster (\code{.tif})}

\item{flm}{character, specifying the file path to a flammability map raster (\code{.tif})}

\item{polyNames}{character, specifying polygon names in \code{summaryPolys} over which
to summarize patch statistics}

\item{summaryPolys}{\code{sf} polygons object}

\item{polyCol}{character, specifying the polygon attribute (column) in \code{summaryPolys} to use}

\item{funList}{list of character strings specifying functions to use for calculating
landscape or patch statistics (e.g., \code{list("lsm_l_area_mn")}).
These functions must work analogously to those in \pkg{landscapemetrics}.
See \code{\link[=default_landscape_metrics]{default_landscape_metrics()}}, \code{\link[=default_patch_metrics]{default_patch_metrics()}}, \code{\link[=default_patch_metrics_seral]{default_patch_metrics_seral()}}.}
}
\description{
Calculate patch statistics (metrics) based on seral stages
}
\seealso{
\code{\link[=patchStats]{patchStats()}}
}
