% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_over_time}
\alias{plot_over_time}
\alias{plot_by_species}
\title{NRV summary plots}
\usage{
plot_over_time(summary_df, ylabel, page = 1)

plot_by_species(summary_df, type = c("box", "violin"), page = 1)
}
\arguments{
\item{summary_df}{summary \code{data.frame} object containing the data to use for plotting}

\item{ylabel}{character, specifying the label for the y-axis}

\item{page}{integer, specifying the \code{facet_wrap_paginate} page to plot.}

\item{type}{character, specifying one of "box" or "violin"}
}
\value{
ggplot object; invoked for side effect of plotting
}
\description{
NRV summary plots
}
