% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seral-stages.R
\name{seralStageMapGeneratorBC}
\alias{seralStageMapGeneratorBC}
\title{Create map of seral stage classes for BC forests}
\usage{
seralStageMapGeneratorBC(cd, pgm, ndtbec)
}
\arguments{
\item{cd}{character, specifying the file path to a \code{cohortData} object (\code{.qs} or \code{.rds})}

\item{pgm}{character, specifying the file path to a \code{pixelGroupMap} raster (\code{.tif})}

\item{ndtbec}{character specifying the file path to e.g., a shapefile delineating
BEC zones and Natural Disturbance Type polygons.}
}
\value{
seral stage map (\code{SpatRaster})
}
\description{
Seral stages based on Table 7 of Cariboo regional Biodiversity Conservation Strategy Report,
and Update Note 3 defining the IDF_Fir and IDF_Pine groups:
}
\details{
paste0("https://www2.gov.bc.ca/assets/gov/farming-natural-resources-and-industry/",
"natural-resource-use/land-water-use/crown-land/land-use-plans-and-objectives/",
"cariboo-region/cariboochilcotin-rlup/bio_strategy_report.pdf")

paste0("https://www2.gov.bc.ca/assets/gov/farming-natural-resources-and-industry/",
"natural-resource-use/land-water-use/crown-land/land-use-plans-and-objectives/"
"cariboo-region/cariboochilcotin-rlup/update_note_3.pdf")
}
\examples{
\dontrun{
if (require("bcdata", quietly = TRUE)) {
  becZones <- bcdata::bcdc_get_data("f358a53b-ffde-4830-a325-a5a03ff672c3")
}
}
}
